Create CodePS5#9
This video was only 30 frames long, so it wasn’t so time-consuming to click on each frame. What if you had hundreds or thousands of frames? Write your own background subtraction code to find the position of the box. Do this by subtracting the first image from each of the images and finding the areas of difference. The particular details of your algorithm may vary; you know the object’s path from watching the movie, and your job is to make Matlab show you a similar path without having you tell Matlab where the object is in each frame (like we did above using ginput). It’s OK if it’s a little noisy, as long as your code makes sense and gives a reasonable estimate of the path. Make a plot of the box’s path with red circles.